requirements.txt
Flask
flask-wtf
flask-sqlalchemy
flask-migrate
flask-login
sendgrid
pyjwt
flask-bootstrap
flask-moment
flask-babel
python-dotenv
rq